refactor(ReciteFragment): clean redundant camera code for readability
